requirements.txt
mysqlclient==1.3.10
psutil==5.2.1
requests==2.13.0
six==1.10.0
tornado==4.4.2
websocket-client==0.40.0
slackclient